further resolve merge conflicts

+++ cpp/src/structure/decompress_to_edgelist_mg_v32_e64.cu
@@ -24,7 +24,7 @@ template std::tuple<rmm::device_uvector<int32_t>,
                     std::optional<rmm::device_uvector<float>>,
                     std::optional<rmm::device_uvector<int64_t>>,
                     std::optional<rmm::device_uvector<int32_t>>>
-decompress_to_edgelist<int32_t, int32_t, float, int32_t, false, true>(
+decompress_to_edgelist<int32_t, int64_t, float, int32_t, false, true>(
   raft::handle_t const& handle,
   graph_view_t<int32_t, int64_t, false, true> const& graph_view,
   std::optional<edge_property_view_t<int64_t, float const*>> edge_weight_view,
@@ -38,7 +38,7 @@ template std::tuple<rmm::device_uvector<int32_t>,
                     std::optional<rmm::device_uvector<float>>,
                     std::optional<rmm::device_uvector<int64_t>>,
                     std::optional<rmm::device_uvector<int32_t>>>
-decompress_to_edgelist<int32_t, int32_t, float, int32_t, true, true>(
+decompress_to_edgelist<int32_t, int64_t, float, int32_t, true, true>(
   raft::handle_t const& handle,
   graph_view_t<int32_t, int64_t, true, true> const& graph_view,
   std::optional<edge_property_view_t<int64_t, float const*>> edge_weight_view,
@@ -52,7 +52,7 @@ template std::tuple<rmm::device_uvector<int32_t>,
                     std::optional<rmm::device_uvector<double>>,
                     std::optional<rmm::device_uvector<int64_t>>,
                     std::optional<rmm::device_uvector<int32_t>>>
-decompress_to_edgelist<int32_t, int32_t, double, int32_t, false, true>(
+decompress_to_edgelist<int32_t, int64_t, double, int32_t, false, true>(
   raft::handle_t const& handle,
   graph_view_t<int32_t, int64_t, false, true> const& graph_view,
   std::optional<edge_property_view_t<int64_t, double const*>> edge_weight_view,
@@ -66,7 +66,7 @@ template std::tuple<rmm::device_uvector<int32_t>,
                     std::optional<rmm::device_uvector<double>>,
                     std::optional<rmm::device_uvector<int64_t>>,
                     std::optional<rmm::device_uvector<int32_t>>>
-decompress_to_edgelist<int32_t, int32_t, double, int32_t, true, true>(
+decompress_to_edgelist<int32_t, int64_t, double, int32_t, true, true>(
   raft::handle_t const& handle,
   graph_view_t<int32_t, int64_t, true, true> const& graph_view,
   std::optional<edge_property_view_t<int64_t, double const*>> edge_weight_view,
